made own service classes to nested properties
