+++ services/ShipService.ts
@@ -1,9 +1,6 @@
-import { CrewMemberFormData, RecordFormData, AllRecords } from './../types/types.d';
 import { Ship } from "../lib/db/entity/Ship";
 import { ShipFormData } from "../types/types";
 import instance from "./instance";
-import { Record } from '../lib/db/entity/Record';
-import { CrewMember } from '../lib/db/entity/CrewMember';
 
 
 class ShipService {
@@ -19,20 +16,5 @@ class ShipService {
         return (await instance.post<Ship>("/ships", formData)).data;
     }
 
-    async addRecord(shipId:string, formData: RecordFormData){
-        return (await instance.post<Record>(`/ships/${shipId}/records`, formData)).data;
-    }
-
-    async getShipRecords(shipId: string) { 
-        return (await instance.get<AllRecords>(`/ships/${shipId}/records`)).data;
-    }
-
-    async getShipRecordByName(shipId: string) { 
-        return (await instance.get<Record[]>(`/ships/${shipId}/records?name=true`)).data;
-    }
-
-    async addCrewMember(shipId:string, formData: CrewMemberFormData){
-        return (await instance.post<CrewMember>(`/ships/${shipId}/crew`, formData)).data;
-    }
 }
 export default new ShipService
